@@ -1,1 +1,9 @@
-<slot></slot><script>import "../app.css";</script><style></style>
+<script>
+	import Footer from '$lib/components/general/footer.svelte';
+	import '../app.css';
+
+	const { children } = $props();
+</script>
+
+{@render children()}
+<Footer />
